Fix test issues and spotbugs.
