--- hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/webapp/QueueCapacityVectorEntryInfo.java
@@ -1,20 +1,0 @@
-package org.apache.hadoop.yarn.server.resourcemanager.webapp;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-
-@XmlRootElement
-@XmlAccessorType(XmlAccessType.FIELD)
-public class QueueCapacityVectorEntryInfo {
-    private String resourceName;
-    private String resourceValue;
-
-    public QueueCapacityVectorEntryInfo() {
-    }
-
-    public QueueCapacityVectorEntryInfo(String resourceName, String resourceValue) {
-        this.resourceName = resourceName;
-        this.resourceValue = resourceValue;
-    }
-}
