+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/webapp/dao/QueueCapacityVectorInfo.java
@@ -1,7 +1,23 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.hadoop.yarn.server.resourcemanager.webapp.dao;
 
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.QueueCapacityVector;
-import org.apache.hadoop.yarn.server.resourcemanager.webapp.QueueCapacityVectorEntryInfo;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -26,5 +42,21 @@ public class QueueCapacityVectorInfo {
             this.capacityVectorEntries.add(new QueueCapacityVectorEntryInfo(
                     queueCapacityVectorEntry.getResourceName(), queueCapacityVectorEntry.getResourceWithPostfix()));
         }
+    }
+
+    public String getConfiguredCapacityVector() {
+        return configuredCapacityVector;
+    }
+
+    public void setConfiguredCapacityVector(String configuredCapacityVector) {
+        this.configuredCapacityVector = configuredCapacityVector;
+    }
+
+    public List<QueueCapacityVectorEntryInfo> getCapacityVectorEntries() {
+        return capacityVectorEntries;
+    }
+
+    public void setCapacityVectorEntries(List<QueueCapacityVectorEntryInfo> capacityVectorEntries) {
+        this.capacityVectorEntries = capacityVectorEntries;
     }
 }
