+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/test/resources/webapp/scheduler-response-PercentageModeLegacyAutoCreation.json
@@ -108,22 +108,6 @@
                     "resourceType" : "COUNTABLE",
                     "units" : "",
                     "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-1",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
-                    "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-2",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
-                    "value" : 0
                   } ]
                 }
               },
@@ -146,22 +130,6 @@
                     "name" : "vcores",
                     "resourceType" : "COUNTABLE",
                     "units" : "",
-                    "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-1",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
-                    "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-2",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
                     "value" : 0
                   } ]
                 }
@@ -619,22 +587,6 @@
                     "resourceType" : "COUNTABLE",
                     "units" : "",
                     "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-1",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
-                    "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-2",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
-                    "value" : 0
                   } ]
                 }
               },
@@ -657,22 +609,6 @@
                     "name" : "vcores",
                     "resourceType" : "COUNTABLE",
                     "units" : "",
-                    "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-1",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
-                    "value" : 0
-                  }, {
-                    "attributes" : { },
-                    "maximumAllocation" : 9223372036854775807,
-                    "minimumAllocation" : 0,
-                    "name" : "custom-resource-2",
-                    "resourceType" : "COUNTABLE",
-                    "units" : "k",
                     "value" : 0
                   } ]
                 }
@@ -935,22 +871,6 @@
                 "resourceType" : "COUNTABLE",
                 "units" : "",
                 "value" : 0
-              }, {
-                "attributes" : { },
-                "maximumAllocation" : 9223372036854775807,
-                "minimumAllocation" : 0,
-                "name" : "custom-resource-1",
-                "resourceType" : "COUNTABLE",
-                "units" : "k",
-                "value" : 0
-              }, {
-                "attributes" : { },
-                "maximumAllocation" : 9223372036854775807,
-                "minimumAllocation" : 0,
-                "name" : "custom-resource-2",
-                "resourceType" : "COUNTABLE",
-                "units" : "k",
-                "value" : 0
               } ]
             }
           },
@@ -973,22 +893,6 @@
                 "name" : "vcores",
                 "resourceType" : "COUNTABLE",
                 "units" : "",
-                "value" : 0
-              }, {
-                "attributes" : { },
-                "maximumAllocation" : 9223372036854775807,
-                "minimumAllocation" : 0,
-                "name" : "custom-resource-1",
-                "resourceType" : "COUNTABLE",
-                "units" : "k",
-                "value" : 0
-              }, {
-                "attributes" : { },
-                "maximumAllocation" : 9223372036854775807,
-                "minimumAllocation" : 0,
-                "name" : "custom-resource-2",
-                "resourceType" : "COUNTABLE",
-                "units" : "k",
                 "value" : 0
               } ]
             }
